+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 * **@dlr-eoc/map-ol:**
  - Use `projjson` or `proj4js` to create projection if defined on `IProjDef`.
+ - Use all `IListMatrixSet` options in `getTileGrid` to create `WMTSTileGrid`.
+ - Allow to pass `olViewOptions` to map-ol component.
+
+* **@dlr-eoc/services-layers:**
+ - More options to create `WMTSTileGrid` on `IListMatrixSet`. The new optional options are `extent?: TGeoExtent`, `origin?: number[]`, `origins?: [number, number][]`, `tileSizes?: number[]` and `sizes?: [number, number][];`.
 
  
 ### Bug Fixes
@@ -14,6 +19,10 @@
 
 * **@dlr-eoc/map-ol:**
  - Transform extent in `setLayerExtentAfterProjection` if nativeBbox.epsg is not new epsg or try to use old extent.
+ - Fix use `l.continuousWorld` for `wrapX` in wmts options.
+
+ * **@dlr-eoc/map-maplibre:**
+ - Allow the usage of different coordinate reference systems in WMTS requests. Use `WmtsLayer.params.matrixSetOptions.matrixSet`. and `WmtsLayer.params.matrixSetOptions.tileMatrixPrefix`
 
 
 # [16.0.0](https://github.com/dlr-eoc/ukis-frontend-libraries/tree/v16.0.0) (2026-05-29) (update dependencies, new library ngx-ukis-ui-clarity and ngx-ukis-utilities, replace core-ui)